scripts/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
httpx>=0.25.0
lxml>=5.0.0
